+++ event_logger_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/xybor-x/xylog"
 )
 
-func TestNewEventLogger(t *testing.T) {
+func TestEventLogger(t *testing.T) {
 	var logger = xylog.GetLogger(t.Name())
 	logger.SetLevel(xylog.DEBUG)
 	xycond.ExpectNotPanic(func() {
@@ -22,4 +22,36 @@ func TestNewEventLogger(t *testing.T) {
 		elogger.Critical()
 		elogger.Log(validCustomLevels[1])
 	}).Test(t)
+}
+
+func TestEventLoggerPair(t *testing.T) {
+	var handler = xylog.NewHandler("", &CapturedEmitter{})
+	handler.SetLevel(xylog.DEBUG)
+	var logger = xylog.GetLogger(t.Name())
+	logger.AddHandler(handler)
+	logger.SetLevel(xylog.DEBUG)
+
+	var elogger = logger.Event("something").
+		Field("foo", "bar").Field("bar", 1).Field("buzzz", true)
+
+	capturedOutput = ""
+	elogger.Debug()
+	xycond.ExpectEqual(
+		capturedOutput, "event=something foo=bar bar=1 buzzz=true").Test(t)
+}
+
+func TestEventLoggerJSON(t *testing.T) {
+	var handler = xylog.NewHandler("", &CapturedEmitter{})
+	handler.SetLevel(xylog.DEBUG)
+	var logger = xylog.GetLogger(t.Name())
+	logger.AddHandler(handler)
+	logger.SetLevel(xylog.DEBUG)
+
+	var elogger = logger.Event("something").
+		Field("foo", "bar").Field("bar", 1).Field("buzzz", true).JSON()
+
+	capturedOutput = ""
+	elogger.Debug()
+	xycond.ExpectEqual(capturedOutput,
+		`{"bar":1,"buzzz":true,"event":"something","foo":"bar"}`).Test(t)
 }