+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Developing
+
+-   Add JSONFormatter.
+-   Change extra, message, and log design.
+
 # v0.0.1 (Sep 02, 2022)
 
 -   Migrated from the [origin project](https://github.com/xybor/xyplatform).